+++ addon/click-on.js
@@ -1,4 +1,5 @@
 import Ember from 'ember';
+import translationOrNull from './utils/translation-or-null';
 
 Ember.Test.registerAsyncHelper('clickOn', function(app, i18n) {
   const {
@@ -6,7 +7,7 @@ Ember.Test.registerAsyncHelper('clickOn', function(app, i18n) {
     t
   } = app.testHelpers;
 
-  const text = t(i18n);
+  const text = translationOrNull(t(i18n)) || i18n;
   const selector = [
     `a:contains("${text}")`,
     `label:contains("${text}")`,