+++ tests/helpers/start-app.js
@@ -3,14 +3,6 @@ import Application from '../../app';
 import config from '../../config/environment';
 import './ember-formulaic/test-helpers';
 
-function setupI18n(application) {
-  const i18n = application.__container__.lookup('service:i18n');
-
-  i18n.on('missing', (locale, key) => {
-    throw new Error(`Missing Translation (${locale}): "${key}"`);
-  });
-}
-
 export default function startApp(attrs) {
   var application;
 
@@ -21,7 +13,6 @@ export default function startApp(attrs) {
     application = Application.create(attributes);
     application.setupForTesting();
     application.injectTestHelpers();
-    setupI18n(application);
   });
 
   return application;